all: rename the module to support v2 versions
Before this change, @latest and specific versions like @v2.4.29 would
fail with module-aware go get, as the module was a v0/v1 instead of the
expected v2.

Amend the go.mod and all the imports. This is just the first step; to
make @latest fully work, we'll need to also release a new tagged
version, and verify that it works.

While at it, update and simplify the 'go get' instructions in the
README.

For #30.